scripts/requirements.txt
python-docx>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
